selenium-agent-mcp-starter-python/requirements.txt
selenium==4.17.2
pytest==8.0.0
python-dotenv==1.0.0
boto3>=1.34,<2
requests>=2.31,<3
hvac>=2.2,<3
webdriver-manager==4.0.1   # optional, for automatic driver management
